+++ test/query-or-where.js
@@ -2,7 +2,7 @@ var assert = require("assert");
 var Query = require("../").Query;
 
 describe("Query.prototype.orWhere()", function () {
-    describe("orWhere(field: string, value: string | number | boolean | Date)", function () {
+    describe("orWhere(field: string, value: any)", function () {
         var query = new Query("users");
 
         query.select("*");
@@ -15,7 +15,7 @@ describe("Query.prototype.orWhere()", function () {
         });
     });
 
-    describe("orWhere(field: string, operator: string, value: string | number | boolean | Date)", function () {
+    describe("orWhere(field: string, operator: string, value: any)", function () {
         var query = new Query("users");
 
         query.select("*");
@@ -28,7 +28,7 @@ describe("Query.prototype.orWhere()", function () {
         });
     });
 
-    describe("orWhere(fields: { [field: string]: string | number | boolean | Date })", function () {
+    describe("orWhere(fields: { [field: string]: any })", function () {
         var query = new Query("users");
 
         query.select("*");